refactored ingestion lambda to contain all funstions in one file due to import issues in aws
