+++ .github/workflows/deploy.yml
@@ -37,6 +37,8 @@ jobs:
           python-version: "3.9"
       - name: Get Make
         run: sudo apt install make -y
+      - name: Install Application Requirements
+        run: make requirements
       - name: Checkout Repo
         uses: actions/checkout@v3
       - name: configure aws credentials
