+++ deploy_ingestion_lambda/src/errors.py
@@ -1,14 +1,14 @@
-class AwsomeError(Exception):
-    pass
+# class AwsomeError(Exception):
+#     pass
 
-class DatabaseConnectionError(AwsomeError):
-    pass
+# class DatabaseConnectionError(AwsomeError):
+#     pass
 
-class IngestionError(AwsomeError):
-    pass
+# class IngestionError(AwsomeError):
+#     pass
 
-class WriteError(AwsomeError):
-    pass
+# class WriteError(AwsomeError):
+#     pass
 
-class SelectQueryError(AwsomeError):
-    pass
+# class SelectQueryError(AwsomeError):
+#     pass
