+++ requirements.txt
@@ -8,7 +8,7 @@ cffi==1.15.1
 charset-normalizer==3.0.1
 click==8.1.3
 coverage==7.1.0
-cryptography==39.0.1
+cryptography==39.0.0
 dparse==0.6.2
 exceptiongroup==1.1.0
 flake8==6.0.0
@@ -47,5 +47,5 @@ toml==0.10.2
 tomli==2.0.1
 types-toml==0.10.8.1
 urllib3==1.26.14
-Werkzeug==2.2.3
+Werkzeug==2.2.2
 xmltodict==0.13.0
